refactor: move Extract and FromRawDict into their own submodule
